Do not embed mapbox/earcut.hpp.
